Add delay, check if done, remove closing window
Login using getpass.
Run the script on a delay of 30 seconds, do this with the time module.
If the image being processed has eyes in the image, add glasses.png over
the eyes.
Append current image to already_done to prevent doing the same image
more than once.
Check width and height of the image once processing.