Reorder karma checking
Reorder the way karma checking is arranged and ordering is done.
Fix broken Twitter auth (due to wrong ordered var assignment.)
Add checking for karma higher than what's in the file. If the current
karma is higher than what is in karma.txt, add it to karma.txt and
update the Twitter profile.